requirements.txt
basecfg~=2.1.1
baselog~=2.0.1
celery[librabbitmq]~=5.4.0
charset_normalizer~=3.3.2
docker~=7.1.0
Flask~=3.0.3
pydantic~=2.7.3
pyyaml~=6.0.1
